Auth validation

+++ pokeapp/views.py
@@ -1,3 +1,6 @@
+from base64 import b64decode
+
+from django.contrib.auth import authenticate
 from django.http import HttpResponse, JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework import status
@@ -6,16 +9,22 @@ from rest_framework.parsers import JSONParser
 
 from pokeapp.models import Trainer
 from pokeapp.serializers import TrainerSerializer
+from .auth import get_or_create_token, get_basic_auth, check_request_token
 
 
 @csrf_exempt
 def trainer_list(request):
+	is_token_valid = check_request_token(request)
+
 	if request.method == 'GET':
 		trainer = Trainer.objects.all()
 		serializer = TrainerSerializer(trainer, many=True)
 		return JsonResponse(serializer.data, safe=False, status=status.HTTP_200_OK)
 
 	elif request.method == 'POST':
+		if not is_token_valid:
+			return HttpResponse(status=status.HTTP_401_UNAUTHORIZED)
+
 		try:
 			data = JSONParser().parse(request)
 		except ParseError:
@@ -32,6 +41,8 @@ def trainer_list(request):
 
 @csrf_exempt
 def trainer_details(request, pk):
+	is_token_valid = check_request_token(request)
+
 	try:
 		trainer = Trainer.objects.get(pk=pk)
 	except Trainer.DoesNotExist:
@@ -42,6 +53,8 @@ def trainer_details(request, pk):
 		return JsonResponse(serializer.data, status=status.HTTP_200_OK)
 
 	elif request.method == 'PUT':
+		if not is_token_valid:
+			return HttpResponse(status=status.HTTP_401_UNAUTHORIZED)
 		try:
 			data = JSONParser().parse(request)
 		except ParseError:
@@ -55,7 +68,23 @@ def trainer_details(request, pk):
 
 
 	elif request.method == 'DELETE':
+		if not is_token_valid:
+			return HttpResponse(status=status.HTTP_401_UNAUTHORIZED)
+
 		trainer.delete()
 		return HttpResponse(status=status.HTTP_204_NO_CONTENT)
 	else:
 		return HttpResponse(status=status.HTTP_405_METHOD_NOT_ALLOWED)
+
+
+
+@csrf_exempt
+def login(request):
+	basic = get_basic_auth(request)
+	if basic is not None:
+		log = b64decode(bytes(basic, 'ascii')).decode('ascii').split(':')
+		user = authenticate(username=log[0], password=log[1])
+		if user is not None:
+			token = get_or_create_token(user)
+			return JsonResponse(data={'token': token.hash})
+	return HttpResponse(status=status.HTTP_400_BAD_REQUEST)
